Fix update-notifier issues
